pyproject.toml
[tool.poetry]
name = "webp-converter"
version = "0.1.0"
description = "webp-converter"
authors = ["KKamJi <xowl5460@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
